requirements.txt
streamlit>=1.35.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
chromadb>=0.5.0
python-dotenv>=1.0.0
pypdf>=4.0.0          # 可选：如需支持PDF文档
jieba>=0.42.1         # 可选：中文分词优化